+++ src/card/cards/control_duration.rs
@@ -8,3 +8,4 @@ pub const WILLOW_SATYR: CardDefinitionId = CardDefinitionId(1474);
 pub const RUBINIA_SOULSINGER: CardDefinitionId = CardDefinitionId(1475);
 pub const CONTROL_MAGIC: CardDefinitionId = CardDefinitionId(1802);
 pub const STEAL_ARTIFACT: CardDefinitionId = CardDefinitionId(1803);
+pub const DISHARMONY: CardDefinitionId = CardDefinitionId(1804);